Initial work on FF4x support.
